runtime/memclr_arm64: change to SIMD memclr
Adapted from the original optimized-routines memset.S at:
https://github.com/ARM-software/optimized-routines/blob/41e5ae54ce6eac3411cd7c51eb40068fbf187601/string/aarch64/memset.S

The memclr code logic is similar to memset, except that the setting
value of memclr is 0, which is not passed by a function argument.
Therefore, there is no need to do the 0-value check below when doing
ZVA.
    TSTW    $255, val
    BNE     no_zva

Another important thing is that, for Go, when count >= 8 bytes, we must
operate with 8-byte alignment to ensure atomic writing of 64-bit
pointers.

Signed-off-by: Xinliang Liu <xinliang.liu@linaro.org>